you guess
A simple while loop enables the user to play a quick guessing game with their computer. The computer will be given a domain of numbers (1,x) from which it will choose a number at random. From there you must guess the number of the computer while it informs you whether your guess was too high or too low.